Move experiments tests around
