+++ app/experimenter/experiments/tests/test_filtersets.py
@@ -1,21 +1,18 @@
 import datetime
-
 import random
-
 from urllib.parse import urlencode
 
-from django.http.request import QueryDict
 from django.conf import settings
-
+from django.http.request import QueryDict
 from django.test import TestCase
 from django.urls import reverse
 
-from experimenter.experiments.models import Experiment
-from experimenter.experiments.tests.factories import ExperimentFactory, ProjectFactory
 from experimenter.experiments.filtersets import ExperimentFilterset
-
+from experimenter.experiments.models import Experiment
+from experimenter.experiments.tests.factories import ExperimentFactory
 from experimenter.experiments.tests.mixins import MockRequestMixin
 from experimenter.openidc.tests.factories import UserFactory
+from experimenter.projects.tests.factories import ProjectFactory
 
 
 class TestExperimentFilterset(MockRequestMixin, TestCase):
